fix Wasi rights system to work with wasi-libc (#6463)
* wasi-common: need FileEntry to track the mode with which a file was opened

* add a test for read, write access modes in path_open

* a directory needs to report full set of rights in fdstat

implementations such as wasi-libc use these as a mask for any rights
they request for a new fd when invoking path_open

* preview2: mask file perms. and we really need to enforce them on stream creation

which will mean editing the wit.

* filesystem.wit: make {read, write, append}-via-stream fallible with an error-code

in order to fail appropriately when a file is not open for reading or
writing.

* filesystem.wit: update comments

* preview2 impls: fix error handling for {read,write,append}_via_stream

* wasi-common: normalize wrong access mode error on windows to badf

* remove kubkob from labeler

I appreciate his work on this stuff in the past but he hasnt been around
for a few years now

* preview 2 filesystem: allow stat on any opened fd

since file perms is now more accurately the file's access mode, a file
open for writing was having problems here.

in reality i dont think it makes sense to restrict this based on the
perms. if a file is opened, you should be able to stat it.

this fixes the host adapter's path_link test, which was broken by prior
changes.  additionally, this fix lets the path_open_dirfd_not_dir test
pass.

* fix the directory base & inheriting rights

in order to work with wasi-testsuite, it needs to be possible to
path_open(dirfd, ".", ...) with the same rights reported in the
fdstat of that dirfd. When we report the Rights::all() set, both
FD_READ and FD_WRITE are set in the base rights, which results in
unix rejecting an openat2(dirfd, ".", O_RDWR) with EISDIR.

By not having the FD_READ and FD_WRITE rights present in the base
rights, the open syscall defaults to O_RDONLY, which is the only
access mode allowed for opening directories.

* path_open of a directory with read and write succeeds on windows