wasi-threads: check module shape at spawn time (#6492)
This change relaxes what kinds of modules can be run when wasi-threads
is enabled via `--wasi-modules experimental-wasi-threads`. Previously,
as reported in #6153, simple modules that made no use of thread spawning
or shared memories were preemptively rejected when the wasi-threads
context was created. This is too restrictive.

Instead, this change does the following:
- it moves the check for whether a module is valid according to the
  wasi-threads specification to the point a new thread is spawned; this
  resolves #6153
- as noted in #6153, this change also adds a better error message
  indicating that wasi-threads expects a shared memory import
- the way this is implemented also improves the module instantiation: by
  constructing an `InstancePre` once when the `WasiThreadsCtx` is built,
  we might shave off a bit of time from the "spawn a thread" call;
  this supercedes a similar effort in #5741