Improve handling of nested continuations (#54)
This PR improves the handling of nested continuations (i.e., invoking continuations within continuations).

It fixes two bugs:
1. In `FuncEnvironment::typed_continuations_load_continuation_object`, we were incorrectly setting the `readonly` flag, which meant that multiple loads to the continuation object slot in the `VMContext` were sometimes combined into one, even though the slot was overwritten in between.
2. In `continuation::resume`, we were not updating the top of stack (TSP) pointer in the `VMInstance` after returning from a continuation.

This PR also adds debug printing to certain functions in `continuation.rs`. Printing only happens in debug builds AND if the variable `ENABLE_DEBUG_PRINTING` in the same file is manually set to true. The printing goes through a macro to ensure that there is no chance of it affecting performance in release builds.

This PR also adds various tests for nested uses of continuations, which also cover the two bugs mentioned earlier.