Consolidate translation of resume instructions in FuncEnvironment (#13)
Currently, the translation of `resume` instructions is somewhat arbitrarily split between `translate_operator` in `code_translator.rs` and `func_environ.rs`. The low-level nature of translating `resume` requires a lot of functionality found only in the latter, which required us to add many helper functions to the `FuncEnvironment` trait (i.e., `spec.rs`). As a result, every change to the translation requires a lot of unnecessary busywork (updating `spec.rs`, `dummy.rs` and `env.rs`).

This particularly obstructs my subsequent optimisation work.

This PR consolidates all of the logic for translating `resume` in `func_environ.rs`. The only work left to do inside  `translate_operator` is very minor, specifically the interaction with the stack and interpreting the resume table. The `FuncTranslationState` is not even handed to the logic inside `func_environ.rs` anymore.

This is purely a refactoring, there are no changes to the actual translation logic.

---------

Co-authored-by: Daniel Hillerström <daniel.hillerstrom@ed.ac.uk>