Move continuation related types into their own crate (#39)
We currently have a situation where types such as `ContinuationObject` and related constants are defined in the crate `wasmtime-runtime`. This means that we cannot access any of our definitions from crates that must not depend on the `runtime` crate (for example, if it causes cyclic dependencies).

This PR remedies this situation by moving the definitions of types and constant from `continuation.rs` in the runtime crate into a new crate, `wasmtime-continuations`. As this crate has no (non-trivial) dependencies, we can use it from everywhere. Note that the runtime crate continues to contain the definitions of the continuation-related libcalls.