Handle duplicated tag entries on `resume`.
The specification allows duplicated occurrences of the same `tag` in
filter list on the `resume` instruction (and `resume_throw`
instruction). Only the first occurrence is relevant as it shadows the
subsequent ones, however, cranelift complains if we emit duplicated
case labels in a `Switch`, e.g.

```
$ ./target/debug/wasmtime wast -W=exceptions,function-references,typed-continuations dup.wast
thread '<unnamed>' panicked at 'Tried to set the same entry 0 twice', cranelift/frontend/src/switch.rs:58:9
```

This patch compiles only the first occurrence of a tag and skips
subsequent entries.