Re-use Payloads object buffers (#47)
Currently, we use `Payloads` objects for example to
- Propagate payloads from `suspend` calls to handlers
- From `resume` calls back to the `suspend` sites

In both cases, we proceed as follows:
1. Allocate a sufficiently sized data buffer within the `Payloads` object first
2. Write the payload data to it
3. Read the payload data on the receiving side
4. Deallocate the buffer within the `Payloads` object.

This PR changes this behaviour: Instead of deallocating the buffer in step 4, we simply mark it as empty, but keep the buffer. Thus, we may change step 1 so that we first check if the `Payloads` object under consideration already has a sufficiently sized buffer. If so, we reuse it, and only need to allocate a new one otherwise.