Use upstream's CI (#81)
This patch replaces the bespoke "WasmFX CI workflow" with the workflows of upstream. The necessary changes to the workflow files are minimal.

However, using their workflows did show that our lockfiles are out of sync quite a bit, which was mainly caused by uses of `num_enum` and `regex` crates. Both uses seem unnecessary to me, however, the debug printing does make internal use of `regex`, and therefore I have disabled it until we have implemented a suitable substitute for the crate.