WasmFX baseline implementation.
This patch adds a baseline implementation for our WasmFX instruction
set implementation. It is a baseline in the sense that it is not
particularly optimised. The implementation is mostly disjoint from the
main (or optimised) implementation, though, there is one particular
code that is shared at the moment and that is
`ContinuationReference`. The current implementation exploits that the
pointer size of `*mut ContinuationObject` and `*mut VMContRef` is the
same. It is a slight hack that probably have to change in the
future. Another important thing to remark is that the implementation
is built on top of vanilla Wasmtime Fiber, meaning that we should in
principle supports the same architectures as Wasmtime Fiber.