WasmFX baseline implementation and modularise implementations (#88)
This patch adds a baseline implementation for our WasmFX instruction
set implementation. It is a baseline in the sense that it is not
particularly optimised. The implementation is mostly disjoint from the
main (or optimised) implementation, though, there is one particular
code that is shared at the moment and that is
`ContinuationReference`. The current implementation exploits that the
pointer size of `*mut ContinuationObject` and `*mut VMContRef` is the
same. It is a slight hack that probably have to change in the
future. Another important thing to remark is that the implementation
is built on top of vanilla Wasmtime Fiber, meaning that we should in
principle support the same architectures as Wasmtime Fiber.

This patch also factors the "main" or "optimised" implementation and
the baseline implementation into separate modules in the cranelift
crate, c.f. the `wasmfx/` subdirectory, where `optimized.rs` contains
the compiler code for the optimised implementation, `baseline.rs`
contains the code for the baseline implementation, and `shared.rs`
contains the common utilities used by both implementations.

Resolves #63.

---------

Co-authored-by: Frank Emrich <git@emrich.io>