Propagate reason for switching stacks through register, remove RunResult (#58)
The API of the original wasmtime-fiber crate parameterises the `Fiber`
struct over types `Resume`, `Return`, and `Yield`. These types reflect
what data is provided when resuming (to the resumed fiber), when the
fiber returns (to the parent) and the data provided when suspending back
to the parent.

We only ever instantiated `Fiber` with `Resume` and `Return` being `()`,
only actually using `Yield` to propagate the tag that we used to
`suspend` with.

Internally, all this propagation is achieved by storing a
`&Cell<RunResult>` on the `FiberStack`. In our case, this means that
propagating the suspended tag involved a pointer indirection.

This PR changes our version of the crate such that
`wasmtime_fibre_switch` may directly propagate an `u64` value through a
register from the caller stack to the stack on which
`wasmtime_fibre_switch` "returns".

We use these `u64` payloads to store values of a new type
`SwitchReason`, which encodes why `wasmtime_fibre_switch` is called. In
particular, it contains the tag index in case we are calling it to
suspend a fiber.

As a result, the `RunResult` enum is not any more and removed, neither
are the type parameters `Resume`, `Return`, and `Yield` of `Fiber`.