Make size of fiber stacks configurable (#114)
This PR adds the flag `wasmfx-stack-size` to the `WasmOptions` group of
`wasmtime` CLI arguments.
This means that values are provided as follows:
```
wasmtime run -Wwasmfx-stack-size=123 ...
```

Its value is used in the implementation of `cont.new` when allocating a
`FiberStack`. The default value of the flag is
`wamtime_runtime::continuation::DEFAULT_FIBER_SIZE` (i..e, the previous
hard-coded value for the fiber stack size).

Inside wasmtime's main configuration struct, of type
`wasmtime::config::Config`, the value of this flag is stored inside a
new struct of type `wasmtime_continuations::WasmFxConfig`. This new
struct is required so that we can use `WasmFxConfig` objects in the
runtime crate, where the overall `wasmtime::config::Config` is not in
scope.

---------

Co-authored-by: Daniel Hillerström <daniel.hillerstrom@ed.ac.uk>