Merge wasmtime-fibre crate into runtime crate (#127)
This turns the `wasmtime_fibre` into a module called `fibre` in the
`runtime` crate. As a result, the `wasmtime_fibre` crate is removed from
the workspace. This is helpful for specializing the fibre implementation
further for our purposes in the future.

Arguably, we could actually make `fibre` a submodule of the existing
`continuation` module in `runtime` (i.e., instead of putting it at
`wasmtime_runtime::fibre`, put it at
`wasmtime_runtime::continuation::fibre`), but I decided against changing
the directory structure further for now.

The only noteworthy changes outside of moving the files are:
1. The types `SwitchDirection` and `SwitchDirectionEnum`, previously in
the `wasmtime_fibre` crate, are now in the `continuations` crate. We
need access to these in cranelift.
2. Instead of adding a lot of /// TODO comments I bailed out of
mandatory doc comments (as imposed by the `runtime` crate) on pub
definitions in the new `fibre` module.