Consolidate translation of wasmfx instructions in FuncEnvironment (#152)
This is just a minor cleanup, influencing where different parts of
translating wasmfx instructions happens.

For `resume` instructions, at some point we moved all of the translation
logic out of `code_translator.rs` and into `func_environ.rs`. This means
that the `Operator::Resume` case in `translate_operator` only
manipulates the value stack, and then delegates the actual translation
to `translate_resume` in `FuncEnvironment` (which then in turn delegates
to the corresponding translation functions in the baseline or optimized
implementation).

This PR does the same for `cont.bind`, `cont.new`, and `suspend`, where
some of the translation logic was scattered between `code_translator.rs`
and `func_environ.rs`. The previous situation had the downside of
requiring a bunch of `typed_continuations_*` functions (e.g.,
`typed_continuations_store_payloads`) to be part of the interface of
`FuncEnvironment`, and then both the baseline and optimized
implementation had to implement that same interface. This created some
unnecessary coupling between the two implementations.

After these changes, the case for any wasmfx instruction `foo` in
`code_translator.rs` only manipulates the value stack, and then calls
`translate_foo` in `FuncEnvironment`, which then calls `translate_foo`
in the baseline or optimized implementation. In particular, no
`typed_continuations_*` functions are required in the `FuncEnvironment`
trait itself anymore.