Add wasmtime_linker_allow_unknown_exports to C API (#173)
The C API for using the wasmtime linker currently only gives access to
the `allow_shadowing` setting of the linker, but not to the
`allow_unknown_exports` setting. However, the latter setting needs to be
enabled in order to allow exporting the global `__stack_pointer`. This
is what's causing the issue described in wasmfx/waeio#2.

(Alternatively, we may add `__stack_pointer` to the list of blessed
globals in the `command` function in
crates/wasmtime/src/runtime/linker.rs, but the comments there suggest
that new items shouldn't be added to that list. Therefore, just making
this setting accessible in the C API seems less controversial.)