Continuation objects as fat pointers (#182)
This patch makes continuation objects allocation free by representing a
continuation object as an immediate, whose 16 high bits are the witness
of the continuation reference revision counter, and the 48 low bits are
the actual continuation reference address. This gives us a cheap way of
implementing the linearity check for continuations. Using 16 bits for
the witness, means that we are currently restricted to 2^16 uses, which
suffices for our examples and benchmarks, but isn't sufficient in
general. However, we can readily generalise this implementation in a
subsequent patch.

The compile time option `unsafe_disable_continuation_linearity_check`
preserves its behaviour.

Resolves #178.