Linking tags (#215)
This patch implements support for importing and exporting tags in the virtual machine. As a consequence tags are now unique to their origin instance, meaning that two different instantiations of the same module yields distinct tags.

The runtime representation of a tag has changed from an `u32` to a `usize` (or native pointer). The identity of a tag is given by its address in the vmcontext.

Resolves https://github.com/wasmfx/wasmfxtime/issues/25.