VMContRef pooling (#221)
This PR extends the existing `WasmFXAllocator` to also allocate
`VMContRefs` in addition to `FiberStacks`. In particular, this means the
*pooling* implementation of the allocator now pools these `VMContRef`
objects, too.

The overall goal is that eventually, when the pooling allocator is
enabled, executing `cont.new` instructions can usually be done without
needing to use the system allocator.

To this end, our pooling allocator is equipped with a `Vec<VMContRef>`.
The underlying mechanism for maintaining which indices into the pool are
free vs used (provided by `SimpleIndexAllocator`) remains unchanged: The
idea is that index `i` in the pool refers to the i-th VMContRef (in the
`Vec`) and the i-th region of stack memory (from the overall `mmap`-ped
stack memory allocation).

One notable aspect is the following: The `allocate` API does not simply
return a `VMContRef` (with a stack attached to it). This is due to how
the setup for `Fiber`'s works in the baseline implementation:
`wasmtime-fiber` dictates that a `Fiber` is obtained by properly
initializing a `FiberStack`, for example by providing the closure to run
inside the `Fiber`. Thus, the creation of the `Fiber` must happen
outside of the pool itself, and remains in the implementation of
`cont_new`.
Instead, the `allocate` function of the `WasmFXAllocator` now returns
two things:
1. A pointer to the `VMContRef`, which currently has no stack attached
to it.
2. A `FiberStack`.

Users of the API can then perform the necessary initializiaton work. For
the baseline implementation, this means creating a `Fiber` from the
`FiberStack` and setting the `fiber` field of the `VMContRef` to it. In
the optimized implementation, it sufficies to call the `initialize`
method of the `FiberStack` and then set the `stack` field of the
`VMContRef`.

A few more implementation details:
- In order to represent a `VMContRef` with no stack attached to it, its
`fiber` field is now `Option`-al in the baseline implementation. In the
optimized implementation, there is now a notion of an unallocated
`FiberStack`, which is used for this purpose.
- In order to initialize the pool's `Vec<VMContRef>`s, we need to be
able to create dummy values of this type. To this end, I've added a
`empty()` function to both `VMContRef` types, which creates an "empty"
(for lack of a better word) `VMContRef`. It's basically a default value,
but it felt weird to actually implement `Default` for `VMContRef`.
- In the optimized implementation when using the pooling allocator, we
leave the `Payloads` objects allocated when putting the `VMContRef`
owning them back into the pool. This means that when we get a
`VMContRef` from the pool, we only need to (re)-allocate `Payloads`
objects if the existing one is too small.