set up Cranelift to allow using stack_switch instruction (#224)
This PR is part of a series that implements native stack switching.

This particular PR sets up the Cranelift configuration used by Wasmtime
so that we may use the `stack_switch` CLIF instruction.

Note that this configuration is operating system specific:
- On Linux, everything is great.
- On Windows, we need to use a different setting, informing Cranelift
that stack switching works slightly different. That *setting* already
exists, even though actually using it would cause a failure within
Cranelift later on. I'm nevertheless performing the proper setup in this
PR because our fork does not compile on Windows to begin with, unless
`wasmfx_baseline` is enabled, in which case none of this matters.
- On macOS, we will currently fail during Engine setup if
`typed-continuations` is enabled. That's because I haven't fully looked
into whether it's safe to just do the same was we are doing on Linux. In
that case, we could actually perform native stack switching on x64 macOS
using the same codegen as for Linux (not that anyone would care).