Update README (#233)
This patch updates the README to reflect the recent changes to the CLI.

I think the following bullet point in the limitations subsection is
outdated:
> - Only a single module can be executed. In particular, providing
additional
> modules using the `--preload` option of `wasmtime` can lead to
unexpected
>  behavior.

I think this should work after the proper tags patch #215. Though, I
haven't tested it yet, so I will leave it be for now.