Split linking_tags.wast into two files (#255)
This PR splits `linking_tags.wast` into two files.

This is a preparation step needed to land fast effect forwarding due to
the following:
`linking_tags.wast` currently has two `assert_suspension` directives. In
the current implementation, we are always on the main stack when we
detect that a tag is unhandled and trap from there. However, in the
presence of fast effect forwarding, we detect the lack of an appropriate
handler at the `suspend` site and trap there (i.e., possibly while still
running on a continuation stack).

As a result, the `assert_suspension` directives in `linking_tags.wast`
trigger the problem described in #253 as soon as fast fowarding lands.
As a workaround, this PR splits the file into two, where the
`assert_suspension` is the very last step in each test (i.e., we avoid
further execution inside the `Store` where we trapped while inside a
continuation). The combined contents of both files is identical to the
original `linking_tags.wast`, but the module `bar` is moved to the end
of the second file.