Fast effect forwarding (#256)
This PR fundamentally changes how we handle tag/effect forwarding:
Currently, at any `suspend` site, we unconditionally attempt to switch
to the parent stack (or fail, if we are already on the the main stack).
In the process, we pass the tag identifier that we suspended with. Then,
we check if the `resume` clause in the parent had a handler for that
tag. If it doesn't, we attempt to suspend again to the parent of the
(original) parent. This process is repeated until we either find a
handler, or reach the main stack (meaning that there was no appropriate
handler).

The new design introduced by this PR works as follows: Every stack (main
stack or continuation) carries a `HandlerList` around. This is
effectively a vector of tag identifiers (i.e., `*mut VMTagDefinition`).

On `resume $ct (on $tag_1 $block_1) ... (on $tag_n $block_n)`, we then
fill the active stack's HandlerList with the following _n_ entries:
```
<address of $tag_1>
...
<address of $tag_n>
```

On `suspend $tag`, we then traverse the chain of active continuations,
inspecting the `HandlerList` of the parents of the currently active
continuation. If we find the address of `$tag` as the _i_th entry of the
handler list of some `resume` instruction in our parents, we directly
switch to the stack that executed the corresponding resume instruction.
In the process, we pass along the list index _i_. Back at the `resume`
site, we can then directly jump to the _i_th handler block, using a jump
table.
If the handler search does not succeed, we immediately trap at the
`suspend` site.

Some implementation notes:
- The search for an appropriate handler is implemented in
`wasmtime_cranelift::wasmfx::optimized::search_handler`, which is used
for emitting code at `suspend` sites.
- While there are a lot of changes to `translate_resume`, it is actually
_simpler_ now! The rather annoying back edge from the forward block to
the resume block is gone, and every block will at most be executed once.
- `ControlEffect` is no longer using `TaggedPointer`. There are simply
no more pointers involved here at all, the only payload to hand from
suspend site to resume site (in addition to the return vs suspend
signal) is the *index* of the handler we want to use.

In general, everything is built to be compatible with `switch` being
added on top of this design: In particular, the `HandlerList` and
aforementioned `search_handler` can easily be adapted to support
`switch`, and the changes to `translate_resume` are made with `switch`
in mind.

---------

Co-authored-by: Daniel Hillerström <daniel.hillerstrom@ed.ac.uk>